refactor: Remove `gen-dex-config` command
Decided to remove this command because the Dex generator just
updates certain keys within the portal configuration map when `dex`
is configured through the `ood_portal.yml` configuration file.

By stripping out Dex, it uncomplicates generating the portal, and
enables updates to the portal to be more atomic as the portal is no
longer required to meet a bunch of pre-conditions.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>