feat: Rename config -> generate
Package name change makes it more clear what the internal package does;
it generates the portal that is then saved to a file.

Also renamed `templates` to `_templates` so that it is ignored by the Go
tool. Templates are still embedded into the generator so that are not required
to be copied in as optional dependencies.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>